Verification of subgroup element for schnorr sigs
When using a non prime order group over an elliptic curve, a point of
the curve can belong to the correct curve but not necessarily to the correct
prime order subgroup subgroup. That commits additionally ensures that this is
the case with the given point if the functionality allows for it.
The scalar must be checked as well that it is in the right range, but currently
group/mod is used which automatically checks that property.